chore: removed kustomizations files
Flux had problem applying the customizations files. A theory is that they were
accidentally commited with a capital K in the filename. Flux tries to autogenerate
kustomization files if there are no present (which Kustomization.yml wasn't
interprated as).
Problem is the autogenerated kustomization files adds all manifests as kubernetes
manifests, which the Kustomization.yaml was interprated as.
Removed Kustomization.yaml files in an attempt to rely on generated
kustomization files instead. If not working, should bring back explicit
kustomization files, but with an lower case K in the filename.